+++ src_cpp/vector_helper.hpp
@@ -1,10 +1,10 @@
 class Vec3{
     public:
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 
-    Vec3(const float inX, const float inY, const float inZ) :
+    Vec3(const double inX, const double inY, const double inZ) :
     x{ inX },
     y{ inY },
     z{ inZ }  {
@@ -19,8 +19,8 @@ class Vec3{
 Vec3 operator-(Vec3 a, Vec3 b);
 Vec3 operator*(Vec3 a, Vec3 b);
 bool Clockwise(Vec3 a, Vec3 b, Vec3 c);
-float Magnitude(Vec3 a); // put inside class
+double Magnitude(Vec3 a); // put inside class
 Vec3 Normalise(Vec3 a); //same 
-Vec3 operator*(Vec3 a, float scale);
+Vec3 operator*(Vec3 a, double scale);
 Vec3 operator+(Vec3 a, Vec3 b);
 Vec3 Intersection(Vec3 dirA, Vec3 dirB, Vec3 posA, Vec3 posB);
